refactor(cli): relocate agent-docs engine to lib/agent-docs
The ~700-line agent-docs generator/injector is shared infrastructure (used by
init, upgrade, the setup-nudge hook, and postinstall), not a command — it lived
in cli/commands/ but was never registered. Move it to lib/agent-docs/ so the API
layer can consume it without importing from cli/, unblocking the upgrade + init
extractions. Pure relocation: importers repointed, no behavior change.

Verified: shape-lock unchanged (no new diff); typecheck:strict + json-api pass;
packages/cli vitest 2073/2073.
