+++ packages/cli/lib/agent-docs/agent-docs.mjs
