+++ packages/cli/lib/agent-docs/agent-docs.path-safety.test.mjs
