+++ packages/cli/lib/agent-docs/agent-docs.test.mjs
